Move UnifiedGPXParser files into separate folder
